qat: drop setupDeviceIDs()
setupDeviceIDs() is obsoleted and the preferred approach is driver_override.

The new_id mechanism was added way before we had the initcontainer support in place.
Furthermore, at least for vfio-pci we don't need it at all if the driver uses
ids=8086:<qat VF dev IDs>.

Drop write attemps to new_id and move the corresponding functionality to
qat-init.sh. With this, we can drop the Apparmor policies from the plugin.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>